Rollback: passing index into data-to-control mappers 92782a87389cc74dd4728977ef04312e3df61f90
Passing this attribute would result in loss of data cohesion if an element was removed from the list state. Because sequence controls were optimized to not need to recreate their entire contents after operations being done on them (append, pop, insert etc.) if an element were to be removed or placed somewhere else than the end of a sequence, the entire sequence would have to be remapped, because the index attribute has shifted for existing entries. It is better then to delegate the task of managing indices in a sequence later down the line where they can be monitored better without compromising performance.